feat: add ability to reset injected answers (#10)
### what
add the ability to clear injected answers


### why

we use `prompt.inject` to inject answers for testing. however, if a test adds too many injections on accident that don't get used by the test, then this leaks over to the next test which which causes test failures. instead, we should be able to reset the injected answers before each test to make sure it's clean

I believe this is what's causing flaky tests in https://linear.app/withgraphite/issue/GT-10743/look-into-why-tests-involving-promptsinject-fail-in-ci