fix: [CLI-273] use negative offset for first() (#15)
### What changed?

Changed the offset in `first()` to reflect that we are moving back by the rendered text length.

### Why?

In the rendering code, cursorOffset is used after rendering the text to move the cursor into the desired position. On first() we were incorrectly using a offset of 0 which would not move the caret. This resulted in the caret being stuck where it was (at the end of the line).

Used the script in the draft PR above to test:

[prompts_testing.mov <span class="graphite__hidden">(uploaded via Graphite)</span> <img class="graphite__hidden" src="https://app.stg.graphite.com/user-attachments/thumbnails/42757caa-0133-4e98-be5a-214e8ca54fe3.mov" />](https://app.stg.graphite.com/user-attachments/video/42757caa-0133-4e98-be5a-214e8ca54fe3.mov)